fix structure to have logic outside of component tsx files. Leave notes to self
